+++ CHANGELOG.md
@@ -5,6 +5,13 @@
         ([#2054](https://github.com/Automattic/pocket-casts-android/pull/2054))
     *   Sleep timer: Start fading out audio when sleep timer has 5 seconds left
         ([#2069](https://github.com/Automattic/pocket-casts-android/pull/2069))
+    *   Sleep timer: Set sleep time end of episode for multiple episodes
+        ([#2097](https://github.com/Automattic/pocket-casts-android/pull/2097))
+*   Updates:
+    *   Improved updating podcast episode when subscribed to thousands podcasts. 
+        ([#2082](https://github.com/Automattic/pocket-casts-android/pull/2082))
+    *   Prioritize embedded chapters over remote ones.
+        ([#2098](https://github.com/Automattic/pocket-casts-android/pull/2098))
 *   Bug Fixes
     *   Fix flashing artwork
         ([#2086](https://github.com/Automattic/pocket-casts-android/pull/2086))
@@ -18,9 +25,6 @@
         ([#2037](https://github.com/Automattic/pocket-casts-android/pull/2037))
     *   Sleep Timer restarts automatically if you play again within 5 minutes
         ([#2048](https://github.com/Automattic/pocket-casts-android/pull/2048))
-*   Updates:
-    *   Updated widget and added two new variants - small and large.
-        ([#2043](https://github.com/Automattic/pocket-casts-android/pull/2043))
 *   Bug Fixes
     *   Improved player view in landscape mode. Added chapter artwork, podcast title, and improved video experience.
         ([#2044](https://github.com/Automattic/pocket-casts-android/pull/1944))